:memo:  chore: move some codes related to review from AccommodationDetail to ReviewSection
